requirements.txt
Django>=5.1.0
ultralytics>=8.3.0
torch>=2.2.0
torchvision>=0.17.0
opencv-python>=4.9.0
onnx>=1.16.0
onnxruntime-gpu>=1.18.0
matplotlib>=3.8.0
scikit-learn>=1.4.0
PyYAML>=6.0.0
